+++ processor/src/test/java/sting/processor/AbstractStingProcessorTest.java
@@ -1,13 +1,11 @@
 package sting.processor;
 
-import com.google.common.collect.ImmutableList;
 import java.util.ArrayList;
 import java.util.List;
 import javax.annotation.Nonnull;
 import javax.annotation.processing.Processor;
-import javax.tools.JavaFileObject;
 import org.realityforge.proton.qa.AbstractProcessorTest;
-import static org.testng.Assert.*;
+import org.realityforge.proton.qa.Compilation;
 
 public abstract class AbstractStingProcessorTest
   extends AbstractProcessorTest
@@ -39,64 +37,46 @@ public abstract class AbstractStingProcessorTest
   @Nonnull
   final String javaOutput( @Nonnull final String classname )
   {
-    return toFilename( "expected", classname, "Sting_", ".java" );
+    return toFilename( classname, "Sting_", ".java" );
   }
 
+  @SuppressWarnings( "SameParameterValue" )
   @Nonnull
   final String graphvizOutput( @Nonnull final String classname )
   {
-    return toFilename( "expected", classname, "", StingProcessor.DOT_SUFFIX );
+    return toFilename( classname, "", StingProcessor.DOT_SUFFIX );
   }
 
   @Nonnull
   final String jsonGraphOutput( @Nonnull final String classname )
   {
-    return toFilename( "expected", classname, "", StingProcessor.GRAPH_SUFFIX );
+    return toFilename( classname, "", StingProcessor.GRAPH_SUFFIX );
   }
 
   @Nonnull
   final String jsonOutput( @Nonnull final String classname )
   {
-    return toFilename( "expected", classname, "", StingProcessor.JSON_SUFFIX );
+    return toFilename( classname, "", StingProcessor.JSON_SUFFIX );
   }
 
-  final void assertDescriptorCount( @Nonnull final ImmutableList<JavaFileObject> output, final long count )
+  @SuppressWarnings( "SameParameterValue" )
+  final void assertJsonDescriptorCount( @Nonnull final Compilation compilation, final long count )
   {
-    assertEquals( output.stream().filter( f -> JavaFileObject.Kind.OTHER == f.getKind() ).count(), count );
+    compilation.assertClassOutputFilenameCount( f -> f.endsWith( ".json" ), count );
   }
 
-  final void assertSourceFileCount( @Nonnull final ImmutableList<JavaFileObject> output, final long count )
+  final void assertJsonDescriptorFile( @Nonnull final Compilation compilation, @Nonnull final String classname )
   {
-    assertEquals( output.stream().filter( f -> JavaFileObject.Kind.SOURCE == f.getKind() ).count(), count );
+    compilation.assertClassOutputFilenamePresent( classname.replace( ".", "/" ) + ".json" );
   }
 
-  final void assertClassFileCount( @Nonnull final ImmutableList<JavaFileObject> output, final long count )
+  final void assertBinaryDescriptorCount( @Nonnull final Compilation compilation, final long count )
   {
-    assertEquals( output.stream().filter( f -> JavaFileObject.Kind.CLASS == f.getKind() ).count(), count );
+    compilation.assertClassOutputFilenameCount( f -> f.endsWith( ".sbf" ), count );
   }
 
-  final void assertJavaFile( @Nonnull final ImmutableList<JavaFileObject> outputs, @Nonnull final String classname )
+  final void assertBinaryDescriptorFile( @Nonnull final Compilation compilation, @Nonnull final String classname )
   {
-    final String filename = "/" + classname.replace( ".", "/" ) + ".java";
-    assertTrue( outputs.stream().anyMatch( f -> JavaFileObject.Kind.SOURCE == f.getKind() &&
-                                                f.getName().endsWith( filename ) ),
-                "Missing java source " + filename );
-  }
-
-  final void assertClassFile( @Nonnull final ImmutableList<JavaFileObject> outputs, @Nonnull final String classname )
-  {
-    final String filename = "/" + classname.replace( ".", "/" ) + ".class";
-    assertTrue( outputs.stream().anyMatch( f -> JavaFileObject.Kind.CLASS == f.getKind() &&
-                                                f.getName().endsWith( filename ) ),
-                "Missing java class " + filename );
-  }
-
-  final void assertDescriptorFile( @Nonnull final ImmutableList<JavaFileObject> outputs,
-                                   @Nonnull final String classname )
-  {
-    final String filename = "/" + classname.replace( ".", "/" ) + ".sbf";
-    assertTrue( outputs.stream()
-                  .anyMatch( f -> JavaFileObject.Kind.OTHER == f.getKind() && f.getName().endsWith( filename ) ),
-                "Missing sting descriptor " + filename );
+    compilation.assertClassOutputFilenamePresent( classname.replace( ".", "/" ) + ".sbf" );
   }
 }