+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Update the `org.realityforge.akasha` artifacts to version `0.30`.
 * Change the format when writing strings to the binary descriptor to to avoid the use of writeUTF which is inefficient.
+* Upgrade the `org.realityforge.proton` artifacts to version `0.57`.
+* Eliminate the depdency on `com.google.testing.compile:compile-testing` to avoid migration complexity when moving to a modern java version.
 * Upgrade the minimum java version to `17`.
 * Change the way that `StingProcessor` checks for explicitly declared constructors to use the public API rather than internal of old JVMs.